Rename JPA test classes from PlatformTCK to JPACDIIntegration
