+++ dev/com.ibm.ws.jpa.tests.jpa_32_fat/fat/src/com/ibm/ws/jpa/FATSuite.java
@@ -23,7 +23,7 @@ import com.ibm.ws.jpa.jpa32.JPABootstrapTest;
 import com.ibm.ws.jpa.jpa32.JakartaDataRecreateTest;
 import com.ibm.ws.jpa.jpa32.JakartaPersistenceDataRecreateTest;
 import com.ibm.ws.jpa.jpa32.JakartaPersistenceTest;
-import com.ibm.ws.jpa.jpa32.PlatformTCKTest;
+import com.ibm.ws.jpa.jpa32.JPACDIIntegrationTest;
 
 import componenttest.containers.TestContainerSuite;
 import componenttest.rules.repeater.RepeatTests;
@@ -35,7 +35,7 @@ import componenttest.topology.database.container.DatabaseContainerFactory;
                 JakartaDataRecreateTest.class,
                 JakartaPersistenceTest.class,
                 JakartaPersistenceDataRecreateTest.class,
-                PlatformTCKTest.class,
+                JPACDIIntegrationTest.class,
                 componenttest.custom.junit.runner.AlwaysPassesTest.class
 })
