+++ dev/com.ibm.ws.jpa.tests.jpa_32_fat/publish/servers/JPACDIIntegrationServer/bootstrap.properties
