+++ dev/com.ibm.ws.jpa.tests.jpa_32_fat/test-applications/jpacdiintegration/resources/jpa-3.2/web/WEB-INF/beans.xml
