+++ dev/com.ibm.ws.jpa.tests.jpa_32_fat/test-applications/jpacdiintegration/src/io/openliberty/jpa/jpacdiintegration/tests/models/TestEntity.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package io.openliberty.jpa.platformtck.tests.models;
+package io.openliberty.jpa.jpacdiintegration.tests.models;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
